Refactor basic components to new folder
