+++ client/src/components/properties/PropertyBasicInfoForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { TextField, MenuItem } from '@mui/material'
 import Grid from '@mui/material/Grid'
-import Card from './Card'
+import Card from '../basic/Card'
 
 interface PropertyBasicInfoFormProps {
   formData: {
