+++ client/src/pages/Properties/CreateProperty.tsx
@@ -3,9 +3,9 @@ import { Box, Alert, CircularProgress } from '@mui/material'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useTheme } from '@mui/material/styles'
 import { useCurrency, CURRENCY_CONFIG } from '../../hooks/useCurrency'
-import Sidebar from '../../components/properties/Sidebar'
-import Titlebar from '../../components/properties/Titlebar'
-import CustomButton from '../../components/properties/CustomButton'
+import Sidebar from '../../components/basic/Sidebar'
+import Titlebar from '../../components/basic/Titlebar'
+import CustomButton from '../../components/basic/CustomButton'
 import ImageUploadSection from '../../components/properties/ImageUploadSection'
 import UnitManagement from '../../components/properties/UnitManagement'
 import PropertyBasicInfoForm from '../../components/properties/PropertyBasicInfoForm'
