+++ client/src/pages/Properties/Property.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react'
 import { Box, Typography, Alert } from '@mui/material'
 import { useNavigate } from 'react-router-dom'
-import Sidebar from '../../components/properties/Sidebar'
-import Titlebar from '../../components/properties/Titlebar'
+import Sidebar from '../../components/basic/Sidebar'
+import Titlebar from '../../components/basic/Titlebar'
 import PropertyCard from '../../components/properties/PropertyCard'
 import PropertyFilters from '../../components/properties/PropertyFilters'
 import ErrorBoundary from '../../components/ErrorBoundary'
